Bug #1695645 unisoc: audio: fix coverity warning
[root cause  ]: coverity static code analysis warning
[changes     ]: unisoc audio
[side effects]: no
[self test   ]: yes
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: sharkl5pro
[test case]: NA
[reviewers   ] gaofeng.zhu
[change_type ] feature_bugfix
[tag_product ] common

+++ drivers/unisoc_platform/audio/sprd/dai/vbc/v4/vbc-dai.c
@@ -2715,7 +2715,7 @@ static int vbc_turning_profile_loading(const u8 *profile_data, size_t profile_si
 	}
 
 	offset = sizeof(struct vbc_fw_header);
-	len = p_profile_setting->hdr[profile_id].num_mode * p_profile_setting->hdr[profile_id].len_mode;
+	len = (size_t)(p_profile_setting->hdr[profile_id].num_mode) * p_profile_setting->hdr[profile_id].len_mode;
 	if (p_profile_setting->data[profile_id] == NULL) {
 		p_profile_setting->data[profile_id] = kzalloc(len, GFP_KERNEL);
 		if (p_profile_setting->data[profile_id] == NULL) {
